requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0
torch>=2.0.0
ta>=0.11.0
pycoingecko>=3.1.0
python-binance
optuna
backtesting>=0.3.3
streamlit>=1.28.0
vectorbt>=0.26.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
python-dotenv>=1.0.0
pyyaml>=6.0
pyarrow>=14.0.0
requests>=2.31.0
tqdm>=4.66.0
joblib>=1.3.0
